+++ backend/src/main/java/com/kozimor/wms/Database/Model/DTO/ItemDTO.java
@@ -23,7 +23,7 @@ public class ItemDTO {
     private String description;
     private String categoryName;
     private UnitType unit;
-    private Integer currentQuantity;
+    private Double currentQuantity;
     private String qrCode;
     private String createdAt;
     private String updatedAt;
