+++ backend/src/main/java/com/kozimor/wms/Database/Model/DTO/TransactionRequestDTO.java
@@ -21,7 +21,7 @@ public class TransactionRequestDTO {
 
     @NotNull(message = "Quantity is required")
     @Min(value = 1, message = "Quantity must be greater than 0")
-    private Integer quantity;
+    private Double quantity;
 
     private Long userId;
