+++ backend/src/main/java/com/kozimor/wms/Database/Model/Order.java
@@ -60,9 +60,9 @@ public class Order {
     @Column(name = "updated_at")
     private LocalDateTime updatedAt;
 
-    public Integer getTotalQuantity() {
+    public Double getTotalQuantity() {
         return orderLines.stream()
-                .mapToInt(OrderLine::getQuantity)
+                .mapToDouble(OrderLine::getQuantity)
                 .sum();
     }
